requirement.txt
asgiref==3.5.2
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
boto3==1.24.7
botocore==1.27.7
colorama==0.4.4
decorator==5.1.1
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
executing==0.8.3
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
jmespath==1.0.0
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.12.0
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.2.2.post1
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
